Feat: Authorization과정 ensureAuthorization함수 생성
1. CartController.js에 적용
2. LikeController.js에 적용
3. OrderController.js에 적용
4. carts.js parameter 값 이름 수정

+++ controller/CartController.js
@@ -1,3 +1,4 @@
+import jwt from 'jsonwebtoken';
 import connection from '../mariadb.js';
 import { StatusCodes } from 'http-status-codes';
 import dotenv from 'dotenv';
@@ -6,10 +7,11 @@ dotenv.config();
 // 장바구니 담기
 export const addToCart = async (req, res) => {
   try {
-    const { book_id, quantity, user_id } = req.body;
+    const { book_id, quantity } = req.body;
 
+    let authorization = ensureAuthorization(req, res);
     let sql = `INSERT INTO cartItems (book_id, quantity, user_id) VALUES (?,?,?)`;
-    let values = [book_id, quantity, user_id];
+    let values = [book_id, quantity, authorization.id];
 
     const [results] = await connection.query(sql, values);
     res.status(StatusCodes.OK).json(results);
@@ -21,12 +23,13 @@ export const addToCart = async (req, res) => {
 // 장바구니 조회  // 선택된 장바구니 아이템 조회
 export const getCartItems = async (req, res) => {
   try {
-    const { user_id, selected } = req.body;
+    const { selected } = req.body;
+    let authorization = ensureAuthorization(req, res);
     let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                 FROM cartItems LEFT JOIN books 
                 ON cartItems.book_id = books.id
                 WHERE user_id=? AND cartItems.id IN (?);`;
-    let values = [user_id, selected];
+    let values = [authorization.id, ...selected];
     const [results] = await connection.query(sql, values);
     res.status(StatusCodes.OK).json(results);
   } catch (err) {
@@ -37,12 +40,35 @@ export const getCartItems = async (req, res) => {
 // 장바구니 도서 삭제
 export const removeCartItem = async (req, res) => {
   try {
-    let { id } = req.params;
-    let sql = `DELETE FROM cartItems WHERE id = ? `;
-    const [results] = await connection.query(sql, id);
-    res.status(StatusCodes.OK).json('장바구니에서 삭제합니당');
+    let { cartId } = req.params;
+    let authorization = ensureAuthorization(req, res);
+    let sql = `DELETE FROM cartItems WHERE id = ? AND  user_id = ?`;
+    let values = [cartId, authorization.id];
+    console.log(values);
+    const [results] = await connection.query(sql, values);
+    // 장바구니에서 없는데도 삭제메세지 출력됨
+    if (results.affectedRows > 0) res.status(StatusCodes.OK).json('식제완료');
+    else res.status(StatusCodes.OK).json('삭제안돼' + results.affectedRows);
   } catch (err) {
     console.error(err);
     res.status(StatusCodes.BAD_REQUEST).end();
   }
 };
+
+// 인증 인가
+function ensureAuthorization(req, res) {
+  try {
+    let receivedJwt = req.headers['authorization'];
+    console.log('received jwt : ', receivedJwt);
+
+    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
+    console.log(decodedJwt);
+    return decodedJwt;
+  } catch (err) {
+    console.error(err);
+    return err;
+    // return res.status(StatusCodes.UNAUTHORIZED).json({
+    //   message: '로그인 세션이 만료 되었다제',
+    // });
+  }
+}
